deployment/requirements.txt
streamlit
pandas
numpy
tensorflow
scikit-learn
joblib
Pillow
opencv-python
scipy